flattened dataset generation procedure - now only 2 files
